Migrate all code to Turbo C and 16-bit real mode
- Declare all local variables at the beginning of functions

- Converts remaining structs not to use anonymous unions

- Don't copy and pass structs as values, as it requires
  built-in functions from Turbo C stdlib

- Make sure to zero-initialize BSS data

- Remove remaining 32-bit boot code

- Convert 32-bit asm code to 16-bit

- Add helper for triggering interrupts

- Add script for generating Makefile rules

- Disable some of the apps until they compile

Lots of work still remaining

+++ include/p_lib.h
@@ -6,7 +6,7 @@ extern int vsnprintf(char *buf, size_t nbyte, const char *fmt, va_list va);
 extern int snprintf(char *buf, size_t nbyte, const char *fmt, ...);
 /* lib/rand.c */
 extern void rand_init(void);
-extern uint32_t rand(void);
+extern uint16_t rand(void);
 /* lib/sleep.c */
 extern void sleep(uint32_t msecs);
 /* lib/string.c */
